pyproject.toml
[tool.poetry]
name = "doccrawler"
version = "0.1.0"
description = ""
authors = ["NintenSAGA <358546067@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
rich = "^12.5.1"
PyYAML = "^6.0"
bs4 = "^0.0.1"
requests = "^2.28.1"
Pillow = "^9.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
